backend/requirements.txt
pydantic
pydantic-settings

langgraph
langchain-openai
langchain-tavily

fastapi
uvicorn